requirements.txt
asgiref             3.11.1
Django              6.0.2
django-querycount   0.8.3
djangorestframework 3.16.1
pip                 24.0
sqlparse            0.5.5